apps/guru-api/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Swiss Ephemeris for astronomical calculations
pyswisseph==2.10.3.2

# HTTP client for external API calls
requests==2.31.0

# Optional: OpenAI client for AI features
openai==1.3.5

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Scheduler dependencies
APScheduler==3.10.4

# Payment gateway dependencies
razorpay==1.4.2
stripe==7.0.0
setuptools>=65.5.0  # Required for pkg_resources used by razorpay

# Notification delivery dependencies
twilio==8.10.0
firebase-admin==6.5.0

